silence compiler warning.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@20521 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseBlame/TortoiseBlame.cpp
@@ -1270,7 +1270,7 @@ void TortoiseBlame::StringExpand(LPWSTR str)
 
 void TortoiseBlame::MakeLower(TCHAR* buffer, size_t len)
 {
-    CharLowerBuff(buffer, len);
+    CharLowerBuff(buffer, (DWORD)len);
 }
 
 void TortoiseBlame::RunCommand(const tstring& command)
